Moved aof and snapshot packages to internal folder
