+++ internal/aof/log/store.go
