+++ internal/aof/preamble/store.go
