+++ internal/snapshot/snapshot.go
